Implement briefify.
Adds information about how many messages we've seen from each
fingerprint to the DB, necessary to properly choose when to
replace old messages with newer ones.

There's definitely some race conditions here if we receive a bunch
of messages from the same fingerprint in very quick succession, but
this shouldn't happen much in practice, and the result is simply
that the probability calculation for replacement is slightly off.